+++ aiarena/core/services/__init__.py
@@ -4,7 +4,6 @@ from .bots import Bots
 from .competitions import Competitions
 from .exceptions import MatchRequestException
 from .ladders import Ladders
-from .maps import Maps
 from .match_requests import MatchRequests
 from .matches import Matches
 from .supporter_benefits import SupporterBenefits
