+++ aiarena/core/services/internal/match_requests.py
@@ -5,8 +5,8 @@ from constance import config
 from aiarena.core.models import Map, Match, Result
 
 from ..exceptions import MatchRequestException
-from ..maps import Maps
 from ..supporter_benefits import SupporterBenefits
+from .maps import Maps
 from .matches import create
